pyproject.toml
[tool.poetry]
name = "otlk"
version = "0.3.0"
description = "転職ドラフト解析用のcliツール"
authors = ["Hiroaki Hasegawa <hasepappa@gmail.com>"]



[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.0.0"
numpy = "^1.18.1"
tqdm = "^4.42.0"
click = "^7.0"
requests = "^2.23.0"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.3.4"
black = "^19.10b0"
flake8 = "^3.7.9"
sphinx = "^2.3.1"
sphinx-rtd-theme = "^0.4.3"
jupyterlab = "^1.2.6"
matplotlib = "^3.1.3"
seaborn = "^0.10.0"
japanize-matplotlib = "^1.0.5"
isort = "^4.3.21"

[tool.poetry.scripts]
otlk="otlk.main:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

